+++ app/components/content/Docs/QRCode/DocsQRCodeGradientScan.vue
@@ -1,0 +1,6 @@
+<template>
+  <div class="relative mx-auto flex aspect-square w-full max-w-60 items-center justify-center">
+    <UiQRCode value="https://uithing.com" size="lg" />
+    <UiQRCodeGradientScan />
+  </div>
+</template>