+++ app/components/content/Docs/QRCode/DocsQRCodeLarge.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="flex items-center justify-center">
+    <UiQRCode value="https://uithing.com" size="lg" />
+  </div>
+</template>